refactor: change the filename of views to presentation
